fix: speed up `just dogfood` execution
Shortly after installation, I ran `just dogfood` in order to gauge the speed of Harper on my machine . It seemed slow, each file taking almost a second to lint. 

The call to `cargo run --release` adds a decent amount of overhead to each file in the loop, so instead we can use `cargo build --release` once before the loop to ensure the binary is built, and then run the binary directly within the loop. After this change, each file lints in ~200ms instead of ~1s