feat(core): split dictionary.dict with a blank line between old vs ours, sort and dedupe ours
"old" is the the traditional dictionary part which I guess came from Hunspell but has changes for our additional affixes and properties.

Then there's one blank line.

"our" is everything we've added to the Hunspell dictionary sorted and dedupded. Sorting helps spot duplicates and helps you see at a glance if a word you want to add is already there.

There was only one dupe: `TODO`  was at both `31417` and `49603`

I've tested that Harper still works as usual with the blank line in the dictionary. But there's a chance it doesn't work with some tool I don't know about.