feat: add `just newest-dict-changes`
It finds the last time in the repo that the curated dictionary was edited and prints out details of all the changes.

You can supply a numeric arg to instead look at previous changes. 1 is the same as no arg.

Feel free to suggest a better name! Is hyphenated or not better?

Ideally a version of this should be in a GitHub Action for visual inspection.

This functionality could also be added to `just addnoun` or an extension of it.

A great improvement would be to call `just getforms` for each added or changed entry to show all the word variants the entry will add to the FST dictionary.