feat: capitalize multi-word country and capital city names
There's some subjectivity about what's regarded as a country.

Some do not fit in either `proper_noun_capitalization.rs` or `matcher.rs`

The former capitalizes every word, but some have one minor word that shouldn't be capitalized. I didn't check if that logic is already built in though.

The latter expects words separated by spaces only, but some use hyphens.

Place names that have an uncapitalized word *and* a hyphen therefor don't fit either. There are some similar cases.

We can use hyphen in the former, but periods don't seem to work for places with names like "St. Foo".

The latter is case sensitive so I put in 3 case variants of each: all lower, first word caps but not the other words, all words caps including minor words which shouldn't be.

For place names that have some other variant such as covering both space and hyphen or with and without accent, this adds up to a lot and it's easy to miss one. Plus any other combination of upper and lower case doesn't get flagged.

A linter to handle the most common of these would be better.