chore: convert one from matcher to phrase_corrections
Is this what you're looking for?
Note that it offers more suggestions per error now.

Maybe we need a new macro that can handle inflected forms?

I don't see another example that handles inflected forms or other kinds of variants to copy.

I'll wait for feedback on this one before proceding with converting the others.