feat: allow blank lines and comments in `dictionary.dict`
Fails on `lints_lots_of_latin_correctly` due to the dictionary containing `et al.` around line 49,839 as the sole dictionary entry containing a space.

Is that actually intentional and supported?

This version would support comments after any whitespace following a dictionary entry with its affix annotation on the same line.

If we do need to support words with spaces then I'll redesign this to require a comment delimiter.