chore: tweak priority/position of statusbar item
We don't have full control over this. I wanted our dialect indicator to be right next to the programming language indicator.

I had it set for the best spot I could get it for VS Code but I mainly use the fork Windsurf, which I couldn't debug the extension in until today.

The new priority is the perfect spot on Windsurf and still pretty good on VS Code.