feat: flag "the my", "my the", etc.
All personal possessive adjectives are handled, but not "its" since that's more often a misspelling or typo for "it's".

This is usually the result of editing but probably sometimes comes from non-native English speakers.

I added the `Clone` trait to `AnyCapitalization` because I think that made a slight optimization or avoided a redundancy, but let me know if I was wrong.

Two edge cases are avoided:

1) If the pronoun follows "the" and is capitalized, it's usually part of a proper noun etc such as "the My Documents folder"

2) "her" is both a possessive adjective and an objec-case pronoun so "her the" is not flagged as in "give her the book"