fix: the `Y` affix flag for `-ly` derives only adverbs.
Previously it was giving all derived words by `adverb` and `adjective` properties. Some adjectives do end in `-ly` such as `friendly` but this is for systematic derivations such as `-ly` changing adjective `happy` to adverb `happily`.

I moved `adverb` to the `base_metadata` section mainly because this is how it's currently done throughout the `affixes.json` file.

Those should probably change to separate single-meaning flags such as `5Y`, but that's better handled as its own issue.