fix: fix logic and improve explanatory comments (#1267)
* fix: fix logic and improve explanatory comments

The logic was broken probably because it was hard to follow.

A false positive with "let in" was being flagged.

I removed the `4` verb flag from `in` in `dictionary.dict` which was due to obsolete senses.

I fixed the broken logic and improved the comments so the logic should be easier to follow in the future.

* chore: `just format`