feat: `just printaffixes` prints letters free for new flags (#1540)
* feat: `just printaffixes` prints letters free for new flags

I've also renamed it to `just annotations` but kept the old name as an alias.

Annotations that have comment fields will also be in the printout now.

* fix: remove comment

* feat: `suggestannotation`

Pass in a term you want to make a new annotation for and it will check if any of the letters are not already used, or are OK to reuse.