feat: add letter-case flags to `WordMetadata`
(Plus other orthography flags.)

These are all for the "case-folded lexemes" in the dictionary, independent of the spelling in the document. So even at the start of a sentence or in all caps you can match words that are capitalized in the dictionary, etc.

Also added a `harper-cli case-variants` command that will print all the words in the dictionary that fold multiple case variants into a single entry.