Fix 1734 (#1761)
* docs: Clarify user dictionary creation time

In #1734 it was noted that it would be a good thing if the documentation
made it clear that the user dictionary will only exist if a user has
added something to it.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

* fix(messages): Swap 'global' for 'user' where seen

Wherever a user of harper could see 'global dictionary' replace that
with 'user dictionary' to make it consistent with the harper-ls verb
and configuration names, etc.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

---------

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>